Add auto_ack option.
This allows the agent to have full control of when the message is ack'ed
(assuming ack is turned on). By default this is set to true but if you
need more control over when to ack the message you can set it to false in
which case you will need to make sure that you ack the message.